WIP migration of DocSidebar
