+++ docs/src/theme/DocSidebar/Desktop/index.tsx
@@ -1,46 +1,14 @@
+import React, {ComponentProps} from 'react';
+import type DesktopType from '@theme/DocSidebar/Desktop';
+import Desktop from '@theme-original/DocSidebar/Desktop';
+import { TrackedNav } from '@objectiv/tracker-react';
 
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
+type Props = ComponentProps<typeof DesktopType>
 
-import React from 'react';
-import clsx from 'clsx';
-import {useThemeConfig} from '@docusaurus/theme-common';
-import Logo from '@theme/Logo';
-import CollapseButton from '@theme/DocSidebar/Desktop/CollapseButton';
-import Content from '@theme/DocSidebar/Desktop/Content';
-import type {Props} from '@theme/DocSidebar/Desktop';
-
-// OBJECTIV
-import { tagNavigation } from '@objectiv/tracker-browser';
-// END OBJECTIV
-
-import styles from './styles.module.css';
-
-function DocSidebarDesktop({path, sidebar, onCollapse, isHidden}: Props) {
-  const {
-    navbar: {hideOnScroll},
-    hideableSidebar,
-  } = useThemeConfig();
-
+export default function DesktopWrapper(props: Props): JSX.Element {
   return (
-    <div
-      // OBJECTIV
-      {...tagNavigation({id: 'docs-sidebar'})}
-      // END OBJECTIV
-      className={clsx(
-        styles.sidebar,
-        hideOnScroll && styles.sidebarWithHideableNavbar,
-        isHidden && styles.sidebarHidden,
-      )}>
-      {hideOnScroll && <Logo tabIndex={-1} className={styles.sidebarLogo} />}
-      <Content path={path} sidebar={sidebar} />
-      {hideableSidebar && <CollapseButton onClick={onCollapse} />}
-    </div>
+    <TrackedNav>
+      <Desktop {...props} />
+    </TrackedNav>
   );
 }
-
-export default React.memo(DocSidebarDesktop);
