--- docs/src/theme/DocSidebarItem/styles.module.css
@@ -1,17 +1,0 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-@media (min-width: 997px) {
-  .menuHtmlItem {
-    padding: var(--ifm-menu-link-padding-vertical)
-      var(--ifm-menu-link-padding-horizontal);
-  }
-}
-
-.menuExternalLink {
-  align-items: center;
-}
